requirements.txt
fastapi
uvicorn
requests
python-multipart
pytesseract
pillow
PyMuPDF